requirements.txt
itsdangerous==2.0.1
flask
flask-oidc2
certifi
pyopenssl